analyse multiple roots + lovely lovely tidy up
